// src/modules/system/chat-input/Administrator/embeds/enums.ts
export enum AvailableEmbedConfiguration {
  MEMBER_JOIN = 0,
  MEMBER_LEAVE = 1,
}

export enum EmbedConfigurationConstants {
  EMBED_COMMAND_OPTION_NAME = 'embed',
  SHOW_SUBCOMMAND_NAME = 'show',
  CONFIGURE_SUBCOMMAND_NAME = 'configure',
  CONFIGURE_CONTINUE = '@embed-config-continue',
  CONFIGURE_CANCEL = '@embed-config-cancel',
  MANAGE_FIELDS_SUBCOMMAND_NAME = 'fields',
  MANAGE_FIELDS_ADD = 'add',
  MANAGE_FIELDS_REMOVE = 'remove',
  MANAGE_FIELDS_LIST = 'list',
  MANAGE_FIELDS_RESET = 'reset',
}